split `EquipmentListTiles` widget
